Move freehand draw control
