+++ app/assets/scripts/components/project/map/freehand-draw-control.js
